==> Analysis_Python/02_Data_Cleaning.ipynb <==
# Data Cleaning

---
Objectives:-
- To identify and handle missing or null values in the dataset.
- To remove duplicate records to ensure data accuracy.
- To correct data types for date and numeric columns.
- To detect and treat extreme outliers affecting analysis.
- To standardize column names for consistency.
- To save a cleaned dataset for further analysis.

---

Import Libraries & Load Raw Data:-

import pandas as pd
import numpy as np

df = pd.read_csv('../Data/01_Raw_Data/Apple_Global_Raw_Data.csv')
df.head()

Check Missing Values:-

df.isnull().sum()

Handle Missing Values:-

Numeric_Cols = df.select_dtypes(include = ["int64", "float64"]).columns

for col in Numeric_Cols:
    df[col] = df[col].fillna(df[col].median())

Categorical_Cols = df.select_dtypes(include=["object"]).columns

for col in Categorical_Cols:
    df[col] = df[col].fillna(df[col].mode()[0])

df.isnull().sum()

Remove Duplicate Rows:-

df.duplicated().sum()
df = df.drop_duplicates()

df.shape

Fix Data Types;-

df["Date"] = pd.to_datetime(df["Date"], errors="coerce")

df.dtypes

Outlier Handling (Using IQR Method):-

def remove_outliers(column):
    Q1 = df[column].quantile(0.25)
    Q3 = df[column].quantile(0.75)
    IQR = Q3 - Q1
    
    lower = Q1 - 1.5 * IQR
    upper = Q3 + 1.5 * IQR
    
    return df[(df[column] >= lower) & (df[column] <= upper)]

# Apply on important numeric columns
outlier_columns = ["Units_Sold", "Revenue_USD_Million", "Average_Price_USD", "Profit_Margin_%"]

for col in outlier_columns:
    df = remove_outliers(col)

df.shape

Standardize Column Names:-

df.columns = df.columns.str.lower().str.replace(" ", "_")

df.columns

Final Clean Dataset Overview:-

df.info()
df.describe()

Save Cleaned Dataset:-

Cleaned_Path = ('../Data/02_Clean_Data/Apple_Global_Clean_Data.csv') 
df.to_csv(Cleaned_Path, index = False)

print("Cleaned Dataset Saved Successfully")

---

Key Insights:-

- Missing values across **numerical and categorical fields** were successfully treated, ensuring no data gaps for analysis.
- Duplicate entries were identified and removed, improving the reliability and accuracy of business metrics.
- The Date column was converted into proper datetime format, enabling time-based trend and forecasting analysis.
- Extreme outliers in **sales, revenue, pricing, and marketing spend** were controlled to prevent distortion in results.
- Column names were standardized into a clean and consistent format, making further coding and analysis more efficient.
- Date column initially loaded as object type instead of datetime.
- The dataset is now fully structured, clean, and ready for feature engineering and exploratory data analysis.

---
